Task
For this ticket, you will need to modify the method genreSearchQuery in moviesDAO.js to allow the following movie search criteria:

genres: finds all movies that match the search genres
The methods for text and cast searches are already implemented:

textSearchQuery : performs a text search in the movies collection
castSearchQuery: finds movies that include any of the specified cast
You just need to construct the query in genreSearchQuery that queries the movies collection by genre. This method should project all document fields.

A text index was created for you when you restored the collections with mongorestore, so these queries will be performant once they are implemented.

Hint

Check the implementation of similar formats of search criteria - the genres query should be similar.