Modify the method getMovies in moviesDAO.js to allow the application to display new pages of movies.
The UI is already asking for infinite scroll, but it's not implemented yet. Only the first 20 movies (the first "page" of movies) appear on the MFlix homepage.

Once this ticket is completed, scrolling to the bottom of the page will load the next 20 movies, or the next page of movies.