For this Ticket, you'll be required to implement all the methods in usersDAO.js that are called by the API endpoints in users.controller.js. Specifically, you'll implement:
getUser
addUser
loginUser
logoutUser
getUserSession
Registering a new user will insert a document into the users collection, and logging in a user will insert a document into the sessions collection.

There is a unique index on the user_id field in sessions, so we can efficiently query on this field.

MFlix Functionality

Once this ticket is completed, users will be able to register for a new account, log in, logout, and delete their account.