For this ticket, you'll be required to complete the command-line script located in the migrations directory of src called movie-last-updated-migration.js.
Things always change, and a requirement has come down that the lastupdated value in each document of the movies collection needs to be stored as an ISODate rather than a String.

Complete the script so it updates the format of lastupdated using a bulk write. You can find the exact Node.js syntax in the docs
(http://mongodb.github.io/node-mongodb-native/3.1/tutorials/crud/#bulkwrite)
.