Modify the configuration of the MongoClient
For this ticket, you'll be required to modify the configuration of the MongoClient to set the maximum size of the connection pool to 50 connections.

The MongoClient is initialized in the src/index.js file. A link to the URI connection settings is included here for your reference.

Testing and Running the Application

Note:

The unit test only has access to DAO methods, but the connection pool size for the MFlix application is set in the index.js file.

However, the connection pool size for the testing environment is set in test/config/mongoEnvironment.js, so you can test your changes there and the unit test will tell you if something is wrong.

When the unit test passes, make sure to update the code in src/index.js so you can retrieve the validation code from the integration test.