In this case, an InvalidId error will be thrown to getMovieByID()
For this ticket, you'll be required to make the API more robust by handling exceptions. Specifically, what would happen should an incorrectly formatted _id be passed to getMovieByID() in moviesDAO.js?

In this case, an InvalidId error will be thrown to getMovieByID(), but the method does not need to return this error. Instead, if this error is thrown, the method should return null.

A try/catch block is already included for you in getMovieByID(). Use the variable e to figure out if the InvalidId error is being thrown, and then return null in this case.

Hint:

When the error e is caught, it has type Error. You might want to convert this to a string.